貸出物の新規登録処理など修正 プルダウンが
