+++ src/main/resources/templates/lendItemAdd.html
@@ -15,7 +15,7 @@
 		<table border="1"
 		th:if="${categoryId} == 1 or
 			${categoryId} == 2 or
-			${catagoryId} == 3 or
+			${categoryId} == 3 or
 			${categoryId} == 4">
 			<tr>
 				<th>種別</th>
